refactor: id 타입 수정

+++ app/src/main/java/com/teamwss/websoso/data/model/CommentEntity.kt
@@ -5,7 +5,7 @@ import com.teamwss.websoso.data.model.FeedEntity.UserEntity
 data class CommentEntity(
     val user: UserEntity,
     val commentContent: String,
-    val commentId: Int,
+    val commentId: Long,
     val createdDate: String,
     val isModified: Boolean,
     val isMyComment: Boolean,
